% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{CChart}
\alias{CChart}
\title{Creates charts}
\usage{
CChart(
  chart.type,
  x,
  small.multiples = FALSE,
  multi.color.series = FALSE,
  font.units = "px",
  annotation.list = NULL,
  signif.show = FALSE,
  signif.column.comparisons = FALSE,
  ...,
  warn.if.no.match = TRUE,
  append.data = FALSE
)
}
\arguments{
\item{chart.type}{The name of plot to create, e.g \code{\link[flipStandardCharts]{Area}}, \code{\link[flipStandardCharts]{Bar}}, \code{\link[flipStandardCharts]{BarPictograph}}, \code{\link[flipStandardCharts]{Bean}}, \code{\link[flipStandardCharts]{Box}}, \code{\link[flipStandardCharts]{Column}}, \code{\link[flipStandardCharts]{Density}}, \code{\link[flipStandardCharts]{Donut}}, \code{\link[flipStandardCharts]{GeographicMap}}, \code{\link[flipStandardCharts]{Heat}}, \code{\link[flipStandardCharts]{Histogram}}, \code{\link[flipStandardCharts]{Line}}, \code{\link[flipStandardCharts]{Palm}}, \code{\link[flipStandardCharts]{Pie}}, \code{\link[flipStandardCharts]{Pyramid}}, \code{\link[flipStandardCharts]{Radar}}, \code{\link[flipStandardCharts]{Scatter}}, \code{\link[flipStandardCharts]{Stream}}, \code{\link[flipStandardCharts]{TimeSeries}}, \code{\link[flipStandardCharts]{Venn}}, \code{\link[flipStandardCharts]{Violin}}. Spaces will automatically be stripped.}

\item{x}{The data to be plotted.}

\item{small.multiples}{Logical; Whether each series should be shown in its own panel. When this is true, parameters from \code{\link[flipStandardCharts]{SmallMultiples}} can be used (e.g. \code{nrows}, \code{x.order}, \code{share.axes}, \code{average.show}, \code{average.color}, \code{panel.title.show}).}

\item{multi.color.series}{Logical; Indicates whether multiple colors will be shown in a Bar or Column chart with a single series. By default this is off and different colors are used to distinguish between different series. However, when chart.type is "Pyramid", then \code{multi.color.series} is always \code{true}.}

\item{font.units}{One of "px" or "pt"}

\item{annotation.list}{a list of annotations to add to the chart, based on statistics in the input data.}

\item{signif.show}{Logical; whether to show significance tests (arrows or column comparisons).}

\item{signif.column.comparisons}{Logical; whether to show comparisons (letters) in data labels.}

\item{...}{Arguments to the function \code{chart.type}. See documentation for specific chart types or see details below.}

\item{warn.if.no.match}{Logical; If TRUE, a warning is shown if any arugments are not matched.}

\item{append.data}{Logical; If TRUE, extra information is appended to the chart object which is used for exporting. These are appended as attributes
\describe{
\item{ChartData}{This is the data passed to the charting functions after tidying by \code{PrepareData}. It is used for both Excel and Powerpoint exporting.}
\item{ChartSettings}{This a named list so that powerpoint chart can use the same visualization settings. Information about the structure of this list is \href{https://wiki.q-researchsoftware.com/wiki/PptChartSettings}{here}.}
\item{ChartWarning}{This attribute is set if the chart cannot be properly exported to Excel. It is set to a string which is shown as a warning message on export in Displayr.}
\item{ChartType}{This attribute is actually always set, even if \code{append.data} if not true. It is the name of the chart type as used in PowerPoint and can vary depending on the visualization settings used.}
}}
}
\value{
A chart object that can be printed. Most often, a plotly object.
}
\description{
Creates charts
}
\details{
Parameters passed to the charting functions are listed below:
\describe{
    \item{\code{type}}{ For chart types \code{Area}, \code{Bar} and \code{Column}, this can be set to \code{Stacked} to show cumulative values.}
    \item{\code{grid.show}}{ Logical; whether to show grid lines.}
    \item{\code{opacity}}{ Opacity of bars as an alpha value (0 to 1).}
    \item{\code{colors}}{ Character; a vector containing one or more colors specified as hex codes.}
    \item{\code{fit.type}}{ Character; type of line of best fit. Can be one of "None", "Linear", "LOESS",
         "Friedman's super smoother" or "Cubic spline". This option is only valid for charts Area, Bar, Column, Line, Scatter, Pyramid.}
    \item{\code{fit.ignore.last}}{ Logical; whether to ignore the last data point in the fit.}
    \item{\code{fit.line.type}}{ Character; One of "solid", "dot", "dash", "dotdash", or length of dash "2px", "5px".}
    \item{\code{fit.line.colors}}{ Character; a vector containing one or more colors specified as hex codes.}
    \item{\code{fit.line.width}}{ Numeric; Line width of line of best fit.}
    \item{\code{fit.line.name}}{ Character; Name of the line of best fit, which will appear in the hovertext.}
    \item{\code{fit.line.opacity}}{ Opacity of trend line as an alpha value (0 to 1).}
    \item{\code{fit.CI.show}}{ Show 95 percent confidence interval.}
    \item{\code{fit.CI.opacity}}{ Opacity of confidence interval ribbon as an alpha value (0 to 1).}
    \item{\code{fit.CI.colors}}{ Character; a vector containing one or more colors specified as hex codes.}
    \item{\code{title}}{ Character; chart title.}
    \item{\code{title.font.family}}{ Character; title font family. Can be "Arial Black",
"Arial", "Comic Sans MS", "Courier New", "Georgia", "Impact",
"Lucida Console", "Lucida Sans Unicode", "Marlett", "Symbol", "Tahoma",
"Times New Roman", "Trebuchet MS", "Verdana", "Webdings".}
    \item{\code{title.font.color}}{ Title font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{title.font.size}}{ Integer; Title font size; default = 10.}
    \item{\code{subtitle}}{ Character.}
    \item{\code{subtitle.font.color}}{ subtitle font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{subtitle.font.family}}{ Character; subtitle font family.}
    \item{\code{subtitle.font.size}}{ Integer; subtitle font size.}
    \item{\code{footer}}{ Character to show below the chart.}
    \item{\code{footer.font.color}}{ footer font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{footer.font.family}}{ Character; footer font family.}
    \item{\code{footer.font.size}}{ Integer; footer font size.}
    \item{\code{footer.wrap}}{ Logical; whether the footer text should be wrapped.}
    \item{\code{footer.wrap.nchar}}{ Number of characters (approximately) in each
line of the footer when \code{footer.wrap} \code{TRUE}.}
    \item{\code{background.fill.color}}{ Background color in character format (e.g. "black").}
    \item{\code{legend.show}}{ Logical; show the legend.}
    \item{\code{legend.wrap}}{ Logical; whether the legend text should be wrapped.}
    \item{\code{legend.wrap.nchar}}{ Number of characters (approximately) in each
line of the legend when \code{legend.wrap} \code{TRUE}.}
    \item{\code{legend.fill.color}}{ Legend fill color as a named color in character format
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}
    \item{\code{legend.fill.opacity}}{ Legend fill opacity as an alpha value (0 to 1).}
    \item{\code{legend.ascending}}{ Logical; TRUE for ascending, FALSE for descending.
By default, we set it to to FALSE if the chart is stacked and TRUE otherwise.}
    \item{\code{legend.border.color}}{ Legend border color as a named color in character
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}
    \item{\code{legend.border.line.width}}{ Width in pixels of the border
around the legend.  0 = no border.}
    \item{\code{legend.position.x}}{ A numeric controlling the position of the legend. Values range from -0.5 (left) to 1.5 (right).}
    \item{\code{legend.position.y}}{ A numeric controlling the position of the legend. Values range from 0 (bottom) to 1 (top).}
    \item{\code{legend.font.color}}{ Legend font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{legend.font.family}}{ Character; legend font family.}
    \item{\code{legend.font.size}}{ Integer; Legend font size.}
    \item{\code{legend.width}}{ Only used in a TimeSeries chart. Integer; Width (in pixels) of the textbox containing the legend.}
    \item{\code{values.title}}{ Character; title of the axis showing the independent values, i.e. y-axis (in a column or area chart) or x-axis (in a bar chart). This will override dim names read from the input table.}
    \item{\code{values.title.font.color}}{ values-axis title font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{values.title.font.family}}{ Character; values-axis title font family.}
    \item{\code{values.title.font.size}}{ Integer; values-axis title font size.}
    \item{\code{values.line.width}}{ values-axis line width in pixels (0 = no line).}
    \item{\code{values.line.color}}{ values-axis line color as a named color in character format.
(e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}
    \item{\code{values.tick.mark.length}}{ Length of tick marks in pixels. Ticks are only shown when \code{y.line.width > 0}.}
    \item{\code{values.bounds.minimum}}{ Minimum of range for plotting; NULL = no manual range set.}
    \item{\code{values.bounds.maximum}}{ Maximum of range for plotting; NULL = no manual range set.}
    \item{\code{values.tick.distance}}{ Distance between tick marks. Requires that \code{values.bounds.minimum} and \code{values.bounds.maximum} have been set.}
    \item{\code{values.number.ticks}}{ Only used in Stream. The total number of ticks on the y-axis.}
    \item{\code{values.zero}}{ Whether the values-axis should include zero. If false, the range will be determined from the input data.}
    \item{\code{values.zero.line.width}}{ Width in pixels of zero line.}
    \item{\code{values.zero.line.color}}{ Color of horizontal zero line as a named.
color in character format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}
    \item{\code{values.data.reversed}}{ Logical; whether to reverse values-axis or not.}
    \item{\code{values.grid.width}}{ Width of values-grid lines in pixels; 0 = no line.}
    \item{\code{values.grid.color}}{ Color of values-grid lines as a named color in character.
format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, maxColorValue = 255)).}
    \item{\code{values.tick.show}}{ Whether to display the values-axis tick labels.}
    \item{\code{values.tick.suffix}}{ values-axis tick label suffix.}
    \item{\code{values.tick.prefix}}{ values-axis tick label prefix.}
    \item{\code{values.tick.format}}{ A string representing a d3 formatting code. See \url{https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}.}
    \item{\code{values.hovertext.format}}{ A string representing a d3 formatting code. See \url{https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}.}
    \item{\code{values.tick.angle}}{ values-axis tick label angle in degrees 90 = vertical; 0 = horizontal.}
    \item{\code{values.tick.font.color}}{ values-axis tick label font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{values.tick.font.family}}{ Character; values-axis tick label font family.}
    \item{\code{values.tick.font.size}}{ Integer; values-axis tick label font size.}
    \item{\code{categories.title}}{ Character; title of the axis showing the dependent values, i.e. x-axis (in a column or area chart) or y-axis (in a bar chart). This will override dim names read from the input table.}
    \item{\code{categories.title.font.color}}{ categories-axis title font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{categories.title.font.family}}{ Character; categories-axis title font family.}
    \item{\code{categories.title.font.size}}{ Integer; categories-axis title font size.}
    \item{\code{categories.line.width}}{ categories-axis line in pixels, 0 = no line.}
    \item{\code{categories.line.color}}{ categories-axis line color as a named color in character format (e.g. "black") or an rgb value (e.g. rgb(0, 0, 0, macategories.olorValue = 255)).}
    \item{\code{categories.tick.marks}}{ Character; whether and where to show tick marks on the categories.axis. Can be "outside", "inside", "none".}
    \item{\code{categories.tick.mark.length}}{ Length of tick marks in pixels.}
    \item{\code{categories.tick.units}}{ Only used in Stream. "Automatic", "Number", "Day", "Month" or "Year".}
    \item{\code{categories.tick.interval}}{ Only used in Stream. The frequency of ticks on the x-axis. Where the data crosses multiple years, re-starts at each year.}
    \item{\code{categories.bounds.minimum}}{ Minimum of range for plotting; NULL = no manual range set.  Must be less than categories.bounds.maximum.}
    \item{\code{categories.bounds.maximum}}{ Maximum of range for plotting; NULL = no manual range set.  Must be greater than categories.bounds.minimum.}
    \item{\code{categories.tick.distance}}{ Tick mark distance in categories.axis units between minimum and maximum for plotting; NULL = no manual range set.}
    \item{\code{categories.zero}}{ Whether the categories-axis should include zero. If false, the range will be determined from the input data.}
    \item{\code{categories.zero.line.width}}{ Width in pixels of zero line.}
    \item{\code{categories.zero.line.color}}{ Color of horizontal zero (origin) line as a named color in character format (e.g. "black") or hex code.}
    \item{\code{categories.data.reversed}}{ Logical; whether to reverse categories-axis or not.}
    \item{\code{categories.grid.width}}{ Width of y-grid lines in pixels; 0 = no line.}
    \item{\code{categories.grid.color}}{ Color of y-grid lines as a named color in character format (e.g. "black") or hex code.}
    \item{\code{categories.tick.show}}{ Whether to display the categories-axis tick labels.}
    \item{\code{categories.tick.suffix}}{ categories-axis tick label suffix.}
    \item{\code{categories.tick.prefix}}{ categories-axis tick label prefix.}
    \item{\code{categories.tick.format}}{ A string representing a d3 formatting code. See \url{https://github.com/mbostock/d3/wiki/Formatting#numbers}.}
    \item{\code{categories.hovertext.format}}{ A string representing a d3 formatting code. See \url{https://github.com/mbostock/d3/wiki/Formatting#numbers}.}
    \item{\code{categories.tick.angle}}{ categories-axis tick label angle in degrees. 90 = vertical; 0 = horizontal.}
    \item{\code{categories.tick.font.color}}{ categories-axis tick label font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{categories.tick.font.family}}{ Character; categories-axis tick label font family.}
    \item{\code{categories.tick.font.size}}{ Integer; categories-axis tick label font size.}
    \item{\code{categories.tick.label.wrap}}{ Logical; whether to wrap long labels on the categories-axis.}
    \item{\code{categories.tick.label.wrap.nchar}}{ Integer; number of characters in each line when \code{label.wrap} is \code{TRUE}.}
    \item{\code{tooltip.show}}{ Logical; whether to show a tooltip on hover.}
    \item{\code{modebar.show}}{ Logical; whether to show the zoom menu buttons or not.}
    \item{\code{global.font.family}}{ Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}
    \item{\code{global.font.color}}{ Global font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{data.label.show}}{ Logical; whether to show data labels.}
    \item{\code{data.label.font.family}}{ Character; font family for data label.}
    \item{\code{data.label.font.size}}{ Integer; Font size for data label.px.}
    \item{\code{data.label.font.color}}{ Font color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{data.label.format}}{ A string representing a d3 formatting code. See \url{https://github.com/mbostock/d3/wiki/Formatting#numbers}.}
    \item{\code{data.label.prefix}}{ Character; prefix for data values.}
    \item{\code{data.label.suffix}}{ Character; suffix for data values.}
    \item{\code{data.label.threshold}}{ The proportion of the total range below which data labels should not be displayed. Only applicable for pie, bar and column charts.}
    \item{\code{margin.top}}{ Margin between plot area and the top of the graphic in pixels.}
    \item{\code{margin.bottom}}{ Margin between plot area and the bottom of the graphic in pixels.}
    \item{\code{margin.left}}{ Margin between plot area and the left of the graphic in pixels.}
    \item{\code{margin.right}}{ Margin between plot area and the right of the graphic in pixels.}
    \item{\code{margin.inner.pad}}{ Padding in pixels between plot proper and axis lines.}
    \item{\code{background.fill.opacity}}{ Background opacity as an alpha value (0 to 1).}
    \item{\code{charting.area.fill.color}}{ Charting area background color as a named color in character format (e.g. "black") or hex code.}
    \item{\code{charting.area.fill.opacity}}{ Charting area background opacity as an alpha value (0 to 1).}
    \item{\code{line.thickness}}{ Only used in Scatter, Line Area or TimeSeries charts. Controls the line thickness in pixels.}
    \item{\code{line.colors}}{ Character; a vctor containing one or more colors specified as hex codes.}
    \item{\code{line.thickness}}{ Opacity of series lines as an alpha value (0 to 1).}
    \item{\code{marker.show}}{ Only used in Scatter, Line or Area charts. Can be "none", "automatic" or a vector referencing the plotly symbol dictionary using either numerics or strings.}
    \item{\code{marker.colors}}{ Only works if \code{marker.show} is true. Character; a vector containing one or more colors specified as hex codes.}
    \item{\code{marker.opacity}}{ Only works if \code{marker.show} is true. Controls ppacity for markers as an alpha value (0 to 1).}
    \item{\code{marker.size}}{ Only works if \code{marker.show} is true. Controls size in pixels of marker.}
    \item{\code{marker.border.width}}{ Width in pixels of border/line around series bars; 0 is no line.}
    \item{\code{marker.border.colors}}{ Character; a vector containing one or more colors specified as hex codes.}
    \item{\code{marker.border.opacity}}{ Opacity of border around bars as an alpha value (0 to 1).}
    \item{\code{bar.gap}}{ Only used in charts Bar, Column and Pyramid. Chart proportion between each bar or column if using bar or column charts, or between each cluster of bars or columns.}
    \item{\code{weights}}{ Only used in Bean, Box, Density, Histogram, Violin plots or Venn diagrams. An optional \code{\link{list}}, where each element is a vector containing weights corresponding to the values of \code{x}, or, a vector where the weights is assumed applicable for each element in \code{x}.}
    \item{\code{vertical}}{ Only used in Bean, Box, Density, Histogram or Violin plots Display the densities vertically.}
    \item{\code{show.mean}}{ Only used in Violin plots. Displays the mean of the data.}
    \item{\code{mean.color}}{ Defaults to "white".}
    \item{\code{show.median}}{ Only used in Violin plots. Displays the median of the data.}
    \item{\code{median.color}}{ Defaults to "black".}
    \item{\code{show.quartiles}}{ Only used in Violin plots. Displays the quartiles of the data.}
    \item{\code{quartile.color}}{ Defaults to "black".}
    \item{\code{show.range}}{ Only used in Violin plots. Displays the range of the data.}
    \item{\code{range.color}}{ Defaults to "black".}
    \item{\code{show.values}}{ Only used in Box, Density or Histogram (always true in Bean plots). Produces a rug plot of individual values.}
    \item{\code{values.color}}{ Defaults to "Green".}
    \item{\code{show.density}}{ Only used in Violin plots. Show the left or top (if \code{vertical} is FALSE) of the violin plot.}
    \item{\code{density.color}}{ Only used in Bean, Box, Density, Histogram or Violin plots. Defaults to "Green".}
    \item{\code{bw}}{ The smoothing bandwidth to be used when creating a Density, Bean, or Violin plot. This defaults to \code{"nrd0"}, whereas \code{"SJ"} may often be superior (see \code{\link{density}}). The default is to \code{"nrd0"} as \code{"SJ"} fails with trivial categorical cases.}
    \item{\code{adjust}}{ A scaling factor for the bandwidth when creating a Density, Bean, or Violin plot. E.g., a value of 0.5 sets the bandwidth to have of that computed using \code{bw}.}
    \item{\code{kernel}}{ The kernel used when when creating a Density, Bean, or Violin plot. One of "gaussian" (the default), "epanechnikov", "rectangular", "triangular", "biweight", "cosine", "optcosine".}
    \item{\code{n}}{ The number of equally-spaced points at which the density is to be estimated when creating a Density, Bean, or Violin plot. If greater than 512, it is rounded to a power of 2 (see \code{link{density}}).}
    \item{\code{from}}{ The left-most point of the grid used when creating a Density, Bean, or Violin plot.}
    \item{\code{to}}{ The right-most point of the grid used when creating a Density, Bean, or Violin plot.}
    \item{\code{cut}}{ By default, the values of \code{from} and \code{to} are \code{cut} bandwidths beyond the extremes of the data.}
    \item{\code{automatic.lower.density}}{ When \code{TRUE}, which is the default, \code{from} is set to the lowest value in the data.}
    \item{\code{histogram.cumulative}}{ Only used in Histogram plots. Plots the cumulative histogram, if set to TRUE.}
    \item{\code{histogram.counts}}{ Only used in Histogram plots. Displays the counts in tooltips of a histogram, rather than the proportions.}
    \item{\code{maximum.bins}}{ Only used in Histograph plots. The maximum number of bins of the histogram. If \code{NULL}, this is generated automatically.}
    \item{\code{box.points}}{ How outliers are displayed boxplots. \code{"All"} plots all the points. \code{"Suspected outliers"} plots points between 1.5 and 3 IQR from the 1st and 3rd quartile with un-filled circles. \code{"Outliers"} does not plot points between 1.5 and 3 IQR from the 1st and 3rd quartiles.}
    \item{\code{country}}{ Only used in GeographicMap. Character string optionally stating the country that the states are from, if \code{map.type} is \code{"states"}.}
    \item{\code{high.resolution}}{ Only used in GeographicMap. Specifically request a high resolution map. Otherwise the resolution of the map is chosen automatically based on the resolution required for the requested countries or regions.}
    \item{\code{show.missing.regions}}{ Only used in GeographicMap. Logical; Whether to plot regions not included in \code{x} with values of \code{NA}. Used only when \code{mapping.package} is \code{"leaflet"}.}
    \item{\code{treat.NA.as.0}}{ Only used in GeographicMap. Plots any \code{NA} values in the data and any geographical entities without data as having a zero value.}
    \item{\code{ocean.color}}{ Only used in GeographicMap. The color used for oceans (or background).}
    \item{\code{mapping.package}}{ Only used in GeographicMap. Either \code{"leaflet"} (better graphics, wider range of
maps) or \code{"plotly"} (faster).}
    \item{\code{background}}{ Only used in GeographicMap. If \code{"mapping.package"} is \code{"leaflet"}, add a background
tile from opensteetmaps.}
    \item{\code{zip.country}}{ Only used in GeographicMap. One of \code{"Automatic"}, \code{"USA"}, \code{"UK"} or \code{"Australia"}. If \code{"Automatic"} an attempt is made to infer the country from the data.}
    \item{\code{sort.rows}}{ Only used in Heat. Whether to sort rows by their averages or link as a dendrogram. Options are \code{"None"}, \code{"Sort by averages (ascending)"}, \code{"Sort by averages (descending)"} and \code{"Dendrogram"}.}
    \item{\code{sort.columns}}{ Only used in Heat. Whether to sort columns by their averages or link as a dendrogram. Options are \code{"None"}, \code{"Sort by averages (ascending)"}, \code{"Sort by averages (descending)"} and \code{"Dendrogram"}.}
    \item{\code{standardization}}{ Only used in Heat. Whether to standardize the shading of rows or columns. Options are \code{"None"}, \code{"Standardize rows"} and \code{"Standardize columns"}.}
    \item{\code{left.columns}}{ Only used in Heat. An optional list of vectors or matrices to be appended to the left of the heatmap.}
    \item{\code{left.column.headings}}{ Only used in Heat. An optional comma separated string containing headings for \code{left.columns}. If not supplied, colnames of the items in \code{left.columns} are used.}
    \item{\code{right.columns}}{ Only used in Heat. An optional list of vectors or matrices to be appended to the right of the heatmap.}
    \item{\code{right.column.headings}}{ Only used in Heat. An optional comma separated string containing headings for
\code{right.columns}. If not supplied, colnames of the items in \code{right.columns} are used.}
    \item{\code{pie.subslice.colors}}{ A vector containing hex value colors for the outer ring of the pie chart. If not supplied will default to the same colors as the inner ring.}
    \item{\code{pie.subslice.colors.repeat}}{  Logical; if, when a grouped pie chart is displayed, the colors of the subslices should repeat by group, or be different throughout; defaults to TRUE.}
    \item{\code{pie.groups.font.family}}{  Character; font family for group labels.}
    \item{\code{pie.groups.font.size}}{ Font size for group labels.}
    \item{\code{pie.groups.font.color}}{ Font color as a named color or hex code.}
    \item{\code{pie.data.threshold}}{ Labels with values smaller than the theshold are not shown.}
    \item{\code{pie.border.color}}{ A single color for space around pie and between segments.}
    \item{\code{pie.inner.radius}}{ The size of the inner radius of pie and donut charts as a proportion out of 100. Defaults to 70.}
    \item{\code{scatter.x.column}}{ When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the x-coordinate data in a scatterplot.}
    \item{\code{scatter.y.column}}{ When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains the y-coordinate data in a scatterplot. This is ignored if \code{scatter.y} is provided instead.}
    \item{\code{scatter.sizes.column}}{ When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.sizes} data. This is ignored if \code{scatter.sizes} is provided instead.}
    \item{\code{scatter.colors.column}}{ When \code{x} is a dataframe or matrix, the index of the column (1-based) which contains \code{scatter.colors} data. This is ignored if \code{scatter.colors} is provided instead.}
    \item{\code{scatter.groups.column}}{ Only used for small multiples (otherwise use \code{scatter.colors.column} with \code{scatter.colors.as.categorical = TRUE}). The index of the column of \code{x} which is used to aggregate the data for small multiples. By default this is the last column in \code{x}.}
    \item{\code{scatter.labels}}{ Optional vector for labelling scatter points. This should be the same length as the number of observations in x and y. If not provided, the rownames of \code{x} will be used instead.}
    \item{\code{scatter.labels.name}}{ Character; Used for labelling subtitles and footers if \code{scatter.labels} is provided}.
    \item{\code{scatter.sizes}}{ Numeric vector determining of the size of each observation. These can alternatively be provided as a column in \code{x}.}
    \item{\code{scatter.sizes.name}}{ Character; Used for labelling footers and legends if \code{scatter.sizes} is provided.}
    \item{\code{scatter.colors}}{ Numeric, character, or categorical vector determining the color of each observation. These can alternatively be provided as a column in \code{x}.}
    \item{\code{scatter.colors.name}}{ Character; Used for labelling footers if \code{scatter.colors} is provided.}
    \item{\code{scatter.colors.as.categorical}}{ Logical; Whether to treat colors as a categorical groups, or a numeric scale..}
    \item{\code{scatter.labels.as.hovertext}}{ Logical; if TRUE, labels are shown has hovers; otherwise, as a labeled scatterplot.}
    \item{\code{scatter.sizes.as.diameter}}{ Whether to show the points with diameter (instead of area, which is the default) proportional to the sizes variable. Cannot be used if \code{scatter.labels.as.hovertext = FALSE}.}
    \item{\code{scatter.max.labels}}{ Integer; the maximum number of labels to show on a Labeled Scatterplot, i.e. when \code{scatter.labels.as.hovertext = FALSE}.}
    \item{\code{trend.lines}}{ Logical; indicating whether to plot trend lines for multiple tables when \code{scatter.labels.as.hovertext = FALSE}.}
    \item{\code{logos}}{ Optional list of images to be used to label scatterplot instead of the row names. This should be input as a comma-seperated list of URLs. Only used when \code{scatter.labels.as.hovertext = FALSE}.}
    \item{\code{logo.size}}{ Numeric controlling the size of the logos.}
    \item{\code{range.bars}}{ Only used in TimeSeries. Logical; whether the data consists of a single series with low, value, high in the columns, or multiple series.}
    \item{\code{window.start}}{ Only used in TimeSeries. The number of days before the end of the data series to start the range selector window.}
    \item{\code{nrows}}{ Only used for small multiples. Integer; Number of rows to arrange the charts in.}
    \item{\code{x.order}}{ Only used for small multiples. A vector containing the list index of the columns in the order which they are to be shown.}
    \item{\code{share.axes}}{ Only used for small multiples. Force range of the plot to be the same across all panels.}
    \item{\code{average.show}}{ Only used for small multiples. Logical; whether to show a second series in each panel containing the data averaged across all series.}
    \item{\code{average.color}}{ Only used for small multiples. The color in which the average series should be displayed.}
    \item{\code{panel.title.show}}{ Only used for small multiples. Logical; whether to show a title for each panel.}
    \item{\code{panel.title.font.family}}{ Only used for small multiples. Font family of panel titles.}
    \item{\code{panel.title.font.color}}{ Only used for small multiples. Font color of panel titles as hex code.}
    \item{\code{panel.title.font.size}}{ Only used for small multiples. Font size of panel titles.}
    \item{\code{panel.title.wrap}}{ Only used for small multiples. Logical; whether the panel title should be wrapped.}
    \item{\code{panel.title.wrap.nchar}}{ Only used for small multiples. Number of characters (approximately) in each line
    of the panel title when \code{panel.title.wordwrap} \code{TRUE}.}
    \item{\code{pad.top}}{ Only used for small multiples. Numeric in [0,1]; Spacing above chart (between panels).}
    \item{\code{pad.bottom}}{ Only used for small multiples. Numeric in [0,1]; Spacing below chart (between panels).}
    \item{\code{pad.left}}{ Only used for small multiples. Numeric in [0,1]; Spacing to the left of chart (between panels).}
    \item{\code{pad.right}}{ Only used for small multiples. Numeric in [0,1]; Spacing to the right chart (between panels).}
}
}
\examples{
x <- matrix(1:21, 7, 3, dimnames = list(letters[1:7], LETTERS[1:3]))
CChart("Column", x, colors = c("red", "green", "blue"), categories.title = "Categories")
CChart("Bar", x, type = "Stacked", colors = grey(1:3/3), categories.title = "Categories")
CChart("Area", x, small.multiples = TRUE,  colors = rainbow(3), categories.title = "Categories")
}
