% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{convertDataFrameIntoJSON}
\alias{convertDataFrameIntoJSON}
\title{convertDataFrameIntoJSON}
\usage{
convertDataFrameIntoJSON(x, nms, weights, n.decimals)
}
\arguments{
\item{x}{The data.frame.}

\item{nms}{The names of the labels}

\item{weights}{Vector of weights}

\item{n.decimals}{number of decimal places.}
}
\description{
Creating JSON-like list from data frane
}
