% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{getChartFunction}
\alias{getChartFunction}
\title{getChartFunction}
\usage{
getChartFunction(type)
}
\arguments{
\item{type}{The type of chart (i.e., function name mainly)
Returns the function that creates the chart}
}
\description{
getChartFunction
}
