% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{getFunctionAndParameters}
\alias{getFunctionAndParameters}
\title{getFunctionNameAndParameters}
\usage{
getFunctionAndParameters(chart.function.name, small.multiples)
}
\arguments{
\item{chart.function.name}{The name of the function used for creating the chart (aka plot).}

\item{small.multiples}{Logical; whether a panel of charts should be created}
}
\value{
A list with the following elements:
\item{\code{chart.function}}{The function}.
\item{\code{parameter.1}}{The first parameter in \code{chart.function}}.
\item{\code{parameter.o}}{The other parameters in \code{chart.function}}.
}
\description{
Gets the function, loading parameters if necessary, and the parameters of the function.
}
