% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedata.R
\name{isScatter}
\alias{isScatter}
\title{#' Get Data for charting
#'
#' Processes form variables to get first non-NULL
#' @param ... named components of data components
#' @details currently recognized components are input.data.table,
#' input.data.tables, input.data.other, pasted, input.data.raw
#' @return the first non-NULL component \code{...} with an added attribute
#' indicating if the data is raw data
#' @noRd
processDataArgs <- function(..., is.pasted = FALSE)
{  #
    args <- list(...)
    non.null.idx <- which(!vapply(args, function(x) is.null(unlist(x)), FALSE))[1]
    if (is.na(non.null.idx))
        stop("no data supplied", call. = FALSE)
    data <- args[[non.null.idx]]
    input.data.raw <- (is.pasted && inherits(data, "list") && isTRUE(data[[2]])) ||
        names(args)[non.null.idx] %in% c("input.data.raw", "input.data.other")
    attr(data, "input.data.raw") <- input.data.raw
    data
}}
\usage{
isScatter(chart.type)
}
\description{
#' Get Data for charting
#'
#' Processes form variables to get first non-NULL
#' @param ... named components of data components
#' @details currently recognized components are input.data.table,
#' input.data.tables, input.data.other, pasted, input.data.raw
#' @return the first non-NULL component \code{...} with an added attribute
#' indicating if the data is raw data
#' @noRd
processDataArgs <- function(..., is.pasted = FALSE)
{  #
    args <- list(...)
    non.null.idx <- which(!vapply(args, function(x) is.null(unlist(x)), FALSE))[1]
    if (is.na(non.null.idx))
        stop("no data supplied", call. = FALSE)
    data <- args[[non.null.idx]]
    input.data.raw <- (is.pasted && inherits(data, "list") && isTRUE(data[[2]])) ||
        names(args)[non.null.idx] %in% c("input.data.raw", "input.data.other")
    attr(data, "input.data.raw") <- input.data.raw
    data
}
}
