% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{loadPackage}
\alias{loadPackage}
\title{loadPackage}
\usage{
loadPackage(chart.type)
}
\arguments{
\item{chart.type}{The name of the function of the chart type to be loaded.}
}
\description{
Loads the package in which a chart is located. Only does this if the chart's function is registered
in this function as having a package requiring loading.
}
