% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberformats.R
\docType{data}
\name{NumberFormats}
\alias{NumberFormats}
\title{We do not change out the first parameter, as there is no good name for it. E.g., height in barplot, y in Chart, x in pie.}
\format{An object of class \code{character} of length 26.}
\usage{
NumberFormats
}
\description{
We do not change out the first parameter, as there is no good name for it. E.g., height in barplot, y in Chart, x in pie.
}
\keyword{datasets}
