% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{parametersEqual}
\alias{parametersEqual}
\title{parametersEqual}
\usage{
parametersEqual(recipient, donor)
}
\arguments{
\item{recipient}{The name of the parameter in the function.}

\item{donor}{The provided name of the parameter.}
}
\description{
Checks if parameters are equal
}
