% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{parametersEqualAfterSubstitution}
\alias{parametersEqualAfterSubstitution}
\title{parametersEqualAfterSubstitution}
\usage{
parametersEqualAfterSubstitution(recipient, donor)
}
\arguments{
\item{recipient}{The name of the parameter in the function.}

\item{donor}{The provided name of the parameter.}
}
\description{
Checks if parameters are equal after substitution of synonyms
}
