% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedata.R
\name{PrepareData}
\alias{PrepareData}
\title{PrepareData}
\usage{
PrepareData(chart.type, subset = TRUE, weights = NULL,
  input.data.table = NULL, input.data.tables = NULL,
  input.data.raw = NULL, input.data.pasted = NULL,
  input.data.other = NULL, data.source = NULL,
  first.aggregate = NULL, scatter.input.columns.order = NULL,
  scatter.mult.yvals = FALSE, group.by.last = FALSE, tidy = TRUE,
  tidy.labels = FALSE, transpose = FALSE, select.rows = NULL,
  first.k.rows = NA, last.k.rows = NA, select.columns = NULL,
  first.k.columns = NA, last.k.columns = NA, auto.order.rows = FALSE,
  sort.rows = FALSE, sort.rows.exclude = c("NET", "SUM", "Total"),
  sort.rows.column = NULL, sort.rows.decreasing = FALSE,
  auto.order.columns = FALSE, sort.columns = FALSE,
  sort.columns.exclude = c("NET", "SUM", "Total"),
  sort.columns.row = NULL, sort.columns.decreasing = FALSE,
  hide.output.threshold = 0, hide.rows.threshold = 0,
  hide.columns.threshold = 0, reverse.rows = FALSE,
  reverse.columns = FALSE, row.names.to.remove = c("NET", "SUM",
  "Total"), column.names.to.remove = c("NET", "SUM", "Total"),
  split = "[;,]", hide.empty.rows.and.columns = TRUE,
  hide.empty.rows = hide.empty.rows.and.columns,
  hide.empty.columns = hide.empty.rows.and.columns,
  as.percentages = FALSE, categorical.as.binary = NULL,
  date.format = "Automatic", show.labels = TRUE, values.title = "")
}
\arguments{
\item{chart.type}{Character; chart type to be plotted.}

\item{subset}{subset An optional vector specifying a subset of
observations to be used in the fitting process, or, the name of
a variable in \code{data}. It may not be an expression.}

\item{weights}{An optional vector of sampling weights, or, the name
of a variable in \code{data}. It may not be an expression.}

\item{input.data.table}{Array; typically a table of some kind,
which is then processed using
\code{\link[flipTables]{AsTidyTabularData}}.}

\item{input.data.tables}{List of array; each component is assumed
to be a Qtable and will be processed using.
\code{\link[flipTables]{AsTidyTabularData}}}

\item{input.data.raw}{List, containing variables or data.frames.}

\item{input.data.pasted}{List of length six; the first component of
which is assumed to be from a user-entered/pasted table; will
be processed by \code{\link{ParseUserEnteredTable}}.}

\item{input.data.other}{A PickAny Multi Q variable.}

\item{data.source}{Where multiple data inputs are provided, a text
string can be provided to disambiguate. Refer to the source
code for a precise understanding of how this works (it is not
obvious and is not likely to be of any use for most cases, so
should usually be left as a \code{NULL}).}

\item{first.aggregate}{Logical; whether or not the input data needs
to be aggregated in this function. A single variable is
tabulated, 2 variables are crosstabbed if \code{group.by.last} is selected,
and otherwise the mean is computed. If \code{input.data.raw} contains
two an 'X' variable and a 'Y' variable in the first two elements of the list,
the data is automatically aggregated and crosstabbed.}

\item{scatter.input.columns.order}{(deprecated) Use \code{scatter.mult.yvals} instead.}

\item{scatter.mult.yvals}{Logical; When \code{chart.type} is "Scatter',
a \code{TRUE} value indicates that columns of input.data.table or input.data.pasted
should be considered multiple series instead of different attributes (default).}

\item{group.by.last}{Logical; \code{TRUE} and \code{first.aggregate} and there is data
in either of \code{input.data.table} or \code{input.data.pasted}, the data is aggregated
using the last variable}

\item{tidy}{Logical; whether or not the input data needs to be
aggregated in this function (e.g., if an x and y variable have
been provided, a contingency table is used to aggregate. This
defaults to \code{TRUE}. It aggressively seeks to turn the data
into a named vector or a matrix using
\code{\link[flipTables]{TidyTabularData}}. This is not applied
when \code{data.input.tables} are provided, or when the chart
type is any of \code{"Scatter"}, \code{"Bean"},
\code{"Histogram"}, \code{"Density"}, \code{"Box"}, or
\code{"Violin"}.}

\item{tidy.labels}{Logical; whether to remove common prefixes from the
labels of the input data.}

\item{transpose}{Logical; should the resulting matrix (of created)
be transposed?}

\item{select.rows}{String; Comma separated list of rows, by name or index
to select from input table. If blank (default), then all rows are selected.}

\item{first.k.rows}{Integer; Number of rows to select from the top of the input table.}

\item{last.k.rows}{Integer; Number of rows to select from the bottom of the input table.}

\item{select.columns}{String; Comma separated list of columns, by name or index
to select from input table. If blank (default), then all columns are selected.}

\item{first.k.columns}{Integer; Number of columns to select from the left of the input table.}

\item{last.k.columns}{Integer; Number of columns to select from the right of the input table.}

\item{auto.order.rows}{Logical; Automatically order rows by correspondence analysis.}

\item{sort.rows}{Logical; whether to sort the rows of the table. This operation is
performed after row selection. (Ignored if \code{auto.order.rows} is true).}

\item{sort.rows.exclude}{String; If \code{sort.rows} is \code{TRUE}, then rows
in \code{sort.rows.exclude} will be excluded from sorting and
appended at the bottom of the table.}

\item{sort.rows.column}{String; If \code{sort.rows} is true, this column
(specified by name or index) is used for sorting the rows. If not specified,
the column with the largest \code{Column n} or the right-most column
will be used for sorting.}

\item{sort.rows.decreasing}{Logical; Whether rows should be sorted in decreasing order.}

\item{auto.order.columns}{Logical; Automatically order columns by correspondence analysis.}

\item{sort.columns}{Logical; whether to sort the columns of the table.
This operation is performed after column selection (Ignored if
\code{auto.order.columns} is true.}

\item{sort.columns.exclude}{String; If \code{sort.columns} is \code{TRUE}, then columns
in \code{sort.columns.exclude} will be excluded from sorting and
appended at the right of the table.'}

\item{sort.columns.row}{String; If \code{sort.columns} is true, this row
(specified by name or index) is used for sorting the columns. If not specified,
the row with the largest \code{n} or the bottom row
will be used for sorting.}

\item{sort.columns.decreasing}{Logical; Whether columns should be sorted in decreasing order.}

\item{hide.output.threshold}{Integer; If sample size ('Column n' or 'n') is provided
then each cell in the input table will be checked to ensure
'n' or 'Column n' is larger than specified threshold, otherwise an error
message is given.}

\item{hide.rows.threshold}{Integer; If sample size ('Column n' or 'n')
is provided, then rows and with sample sizes smaller than threshold
will be removed from table. Vectors will be treated as 1-d matrices}

\item{hide.columns.threshold}{Integer; If sample size ('Column n' or 'n')
is provided, then columns with sample sizes smaller than threshold
will be removed from table. Vectors will not be affected.}

\item{reverse.rows}{Logical; Whether to reverse order of rows. This operation is
performed after row selection and sorting.}

\item{reverse.columns}{Logical; Whether to reverse order of columns. This operation
is peformed after column selection and sorting.}

\item{row.names.to.remove}{Character vector or delimited string of
row labels specifying rows to remove from the returned table;
default is \code{c("NET", "SUM")}}

\item{column.names.to.remove}{Character vector or delimited string
of column labels specifying columns to remove from the returned
table; default is \code{c("NET", "SUM")}.}

\item{split}{Character delimiter to split
\code{row.names.to.remove} and \code{col.names.to.remove}
on. Default is to split on either of \code{","} or \code{";"}.
Assumed to be a regular expression; see \code{\link{strsplit}}.}

\item{hide.empty.rows.and.columns}{Logical; if \code{TRUE} empty
rows and columns will be removed from the data.  Empty here
meaning that a row or column contains all \code{NA} values, or
in the case of percentages, that a row or column contains only
0's. Retained for backwards-compatibility but is superseded by
\code{hide.empty.rows} and \code{hide.empty.columns}.}

\item{hide.empty.rows}{Logical; hide rows with only NAs or 0's (percentages).}

\item{hide.empty.columns}{Logical; hide columns with only NAs or 0's (percentages).}

\item{as.percentages}{Logical; If \code{TRUE}, aggregate values in the
output table are given as percentages summing to 100. If \code{FALSE},
column sums are given.}

\item{categorical.as.binary}{If data is aggregated and this is true, then categorical variables will be converted into indicator variables for each level in the factor.}

\item{date.format}{One of \code{"Automatic", "US", "International" or "No date formatting"}.
This is used to determine whether strings which are interpreted as dates
in the (row)names will be read in the US (month-day-year) or the
International (day-month-year) format. By default US format is used
if it cannot be deduced from the input data.}

\item{show.labels}{Logical; If \code{TRUE}, labels are used for
names in the data output if raw data is supplied.}

\item{values.title}{The title for the values axis of a chart (e.g.,
the y-axis of a column chart or the x-axis of a bar chart).}
}
\value{
A list with components \itemize{ \item \code{data} - If
    possible, a named vector or matrix, or if that is not posible
    or a data.frame is requested, a data.frame.  \item
    \code{weights} - Numeric vector of user-supplied weights.
    \item \code{values.title} - Character string to be used for the
    y-axis title; will only be a non-empty string if some
    aggregation has been performed on \code{data} \item
    \code{scatter.variable.indices} A named vector indicating which
    columns in \code{data} should be plotted in a scatterplot as
    \code{x}, \code{y}, \code{sizes}, and \code{colors}. Is
    \code{NULL} if \code{chart.type} does not contain
    \code{"Scatter"} or \code{"Bubble"}. \code{NA} is used when the
    data does not exist.  }
}
\description{
Prepares input data for charting.
}
\details{
It is assumed that only one of \code{input.data.pasted},
    \code{input.data.table}, \code{input.data.tables},
    \code{input.data.other}, \code{input.data.raw} is non-NULL.
    They are checked for nullity in that order.
}
\seealso{
\code{\link[flipTables]{AsTidyTabularData}},
    \code{\link[flipData]{TidyRawData}},
    \code{\link[flipTransformations]{ParseUserEnteredTable}}
}
