% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparenumbers.R
\name{PrepareNumbers}
\alias{PrepareNumbers}
\title{Generates all d3 formats for use by CChart.}
\usage{
PrepareNumbers(categories.format.list = NULL, values.format.list = NULL,
  hover.format.list = NULL, data.labels.format.list = NULL,
  as.percentages = FALSE)
}
\arguments{
\item{categories.format.list}{A list of five unnamed items for formatting the categories-axis as
described in \code{\link[flipChartBasics]{ChartNumberFormat}}.}

\item{values.format.list}{As per \code{categories.format.list} except for values-axis formatting.}

\item{hover.format.list}{As per \code{categories.format.list} except for hovertext formatting.}

\item{data.labels.format.list}{As per \code{categories.format.list} except for data label formatting.}

\item{as.percentages}{Whether the default formatting should be as percentages.}
}
\description{
See \url{https://github.com/d3/d3-format} for more information on d3.
Wrapper for \code{ChartNumberFormat}.
}
