% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{recipientIsSubordinateOrEqual}
\alias{recipientIsSubordinateOrEqual}
\title{recipientIsSubordinateOrEqual}
\usage{
recipientIsSubordinateOrEqual(recipient, donor)
}
\arguments{
\item{recipient}{The name of the parameter in the function.}

\item{donor}{The provided name of the parameter.}
}
\description{
All the sub.parts of the donor are in the recipient
}
