% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{recipientIsSubordinateOrEqualAfterSubstitution}
\alias{recipientIsSubordinateOrEqualAfterSubstitution}
\title{recipientIsSubordinateOrEqualAfterSubstitution}
\usage{
recipientIsSubordinateOrEqualAfterSubstitution(recipient, donor)
}
\arguments{
\item{recipient}{The name of the parameter in the function.}

\item{donor}{The provided name of the parameter.}
}
\description{
All the sub.parts of the donor are in the recipient
}
