% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{replaceChartSynonyms}
\alias{replaceChartSynonyms}
\title{replaceChartSynonyms}
\usage{
replaceChartSynonyms(type)
}
\arguments{
\item{type}{The name of the chart type.}
}
\description{
replaceChartSynonyms
}
