% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{RGUIControls}
\alias{RGUIControls}
\title{RGUIControls}
\usage{
RGUIControls(chart.function.name, vector = FALSE, matrix = FALSE,
  raw.data.1 = FALSE, raw.data.2 = FALSE, raw.data.multi = FALSE,
  r.object = FALSE, scalar = FALSE)
}
\arguments{
\item{chart.function.name}{The name of the function used for creating the chart (aka plot).}

\item{vector}{\code{TRUE} if the function accepts a vector as the sole data input.}

\item{matrix}{\code{TRUE} if the function accepts a \code{matrix} as the sole data input.}

\item{raw.data.1}{\code{TRUE} if the function accepts a single variable of 'raw' (non-aggregated) data as an input.}

\item{raw.data.2}{\code{TRUE} if the function accepts a pair of variables of 'raw' data.}

\item{raw.data.multi}{\code{TRUE} if the function accepts multiple variables of 'raw' data as an input.}
}
\value{
A \code{character} object of JavaScript code.
}
\description{
Writes the JavaScript to create the RGUI Controls for Displayr and Q. Note that the first parameter of the function
is not automatically written (as it is assumed to be data, to be  addressed by the remaining arguments).
}
