% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{scaleFontSizes}
\alias{scaleFontSizes}
\title{scaleFontSizes}
\usage{
scaleFontSizes(arguments)
}
\arguments{
\item{arguments}{List of arguments supplied by user}
}
\description{
Convert font size from point to pixel.
}
\details{
All of the charts in flipStandardChart
take font sizes to be in units of pixels, however, textboxes in Displayr
assumes font sizes are in units of points. This function iterates through
\code{user.args} and scales all font sizes so they are in the equivalent points.
Note that arguments not specified by the user are not affected.
}
