% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stacked.cola.associations}
\alias{stacked.cola.associations}
\title{stacked.cola.associations}
\format{
A \code{\link{data.frame}}
}
\source{
Study by Displayr Australia Pty Ltd
}
\description{
Survey responses rating six brands of Cola against Personality traits
}
