% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{Streamgraph}
\alias{Streamgraph}
\title{Streamgraph}
\usage{
Streamgraph(x, colors = c("#5C9AD3", "#ED7D31", "#A5A5A5", "#FFC000",
  "#4473C5", "#70AD46", "#255F91", "#9E480D", "#636365", "#987300", "#26408B",
  "#42682B"), y.axis.show = TRUE, y.number.ticks = 5, hover.decimals = 2,
  x.tick.interval = 1, x.number.format = "\%d \%b \%y", margin.top = 20,
  margin.left = 50, margin.bottom = 30, margin.right = 40)
}
\arguments{
\item{x}{A BasicTable, with columns containing the dates or other numericx-axis variable.}

\item{colors}{The colors of the streams.}

\item{y.axis.show}{If FALSE, the y-axis is not shown.}

\item{y.number.ticks}{The total number of ticks on the y-axis.}

\item{hover.decimals}{The number of decimals to show in hovers.}

\item{x.tick.interval}{The frequency of ticks on the x-axis. Where the data crosses multiple years, re-starts at each year.}

\item{x.number.format}{THe number format of the x-axis. Only integers ('Number') and dates are supported.}

\item{margin.top}{Top margin (default should be fine, this allows for fine-tuning plot space)}

\item{margin.left}{Left margin (default should be fine, this allows for fine-tuning plot space)}

\item{margin.bottom}{Bottom margin (default should be fine, this allows for fine-tuning plot space)}

\item{margin.right}{Right margin (default should be fine, this allows for fine-tuning plot space)}
}
\description{
Area chart centerd around the x-axis rather than on top of it.
}
