% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{substituteArgumentNames}
\alias{substituteArgumentNames}
\title{substituteArgumentNames}
\usage{
substituteArgumentNames(parameter.names, arguments,
  warn.if.no.match = TRUE)
}
\arguments{
\item{parameter.names}{The names of parameters.}

\item{arguments}{The arguments to match to the parameters.}

\item{warn.if.no.match}{If TRUE, a warning is shown if any arugments are not matched.}
}
\description{
substituteArgumentNames
}
