% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{substituteAxisNames}
\alias{substituteAxisNames}
\title{substituteAxisNames}
\usage{
substituteAxisNames(chart.function, arguments)
}
\arguments{
\item{chart.function}{Name of charting function}

\item{arguments}{List of arguments supplied by user}
}
\description{
Substitutes 'categories' or 'values' for 'x' and 'y'.
}
\details{
Unlike \code{substituteArgumentNames}, the behaviour is specific to
  \code{chart.function}. It is also less agggressive in that it searches for
  replacements only at the beginning of the name.
}
