% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\docType{data}
\name{synonyms}
\alias{synonyms}
\title{We do not change out the first parameter, as there is no good name for it. E.g., height in barplot, y in Chart, x in pie.}
\format{An object of class \code{list} of length 7.}
\usage{
synonyms
}
\description{
We do not change out the first parameter, as there is no good name for it. E.g., height in barplot, y in Chart, x in pie.
}
\keyword{datasets}
