% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{CChart}
\alias{CChart}
\title{CChart}
\usage{
CChart(chart.type, x, ..., return.data = FALSE, warn.if.no.match = TRUE,
  append.data = FALSE, series.stack = FALSE,
  scatter.labels.as.hovertext = TRUE)
}
\arguments{
\item{chart.type}{The name of plot to create, e.g "100\% Stacked Column Chart"}

\item{x}{The data to be plotted.}

\item{...}{Arguments to the function \code{chart.type}}

\item{return.data}{Logical; If TRUE, returns the data x instead of plotting the chart.}

\item{warn.if.no.match}{Logical; If TRUE, a warning is shown if any arugments are not matched.}

\item{append.data}{Logical; If TRUE, appends the chart data as an attribute called "ChartData".}

\item{series.stack}{Logical; Option to show charts with series stacked, in
Column, Bar, or Area chart.}

\item{scatter.labels.as.hovertext}{Option to use ScatterChart instead of LabeledScatterChart}
}
\value{
A chart object that can be printed. Most often, a plotly object.
}
\description{
Creates charts
}
\details{
Where \code{chart.type} is not the name of an existing function. It is
  always assumed that the first parameter
  in the signature is a data object, which is assigned the value of \code{x}.
}
