% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{CChart}
\alias{CChart}
\title{CChart}
\usage{
CChart(chart.type, x, ..., warn.if.no.match = TRUE)
}
\arguments{
\item{chart.type}{The name of the function, used for creating the chart (aka plot).}

\item{x}{The data to be plotted.}

\item{...}{Arguments to the function \code{chart.type}}

\item{warn.if.no.match}{If TRUE, a warning is shown if any arugments are not matched.}
}
\description{
Creates charts
}
\details{
Where \code{chart.type} is not the name of an existing function, ea c
}
