% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedata.R
\name{PrepareData}
\alias{PrepareData}
\title{PrepareData}
\usage{
PrepareData(chart.type, subset = TRUE, weights = NULL,
  input.data.table = NULL, input.data.tables = NULL,
  input.data.raw = NULL, input.data.pasted = NULL,
  input.data.other = NULL, first.aggregate = FALSE, tidy = TRUE,
  transpose = FALSE, missing = "Exclude cases with missing data",
  row.names.to.remove = c("NET", "SUM"), column.names.to.remove = c("NET",
  "SUM"), split = "[;,]", as.percentages = FALSE, show.labels = TRUE)
}
\arguments{
\item{chart.type}{Character; chart type to be plotted.}

\item{subset}{subset An optional vector specifying a subset of
observations to be used in the fitting process, or, the name of
a variable in \code{data}. It may not be an expression.}

\item{weights}{An optional vector of sampling weights, or, the name
of a variable in \code{data}. It may not be an expression.}

\item{input.data.table}{Array; typically a table of some kind, which is then
processed using \code{\link[flipTables]{AsTidyTabularData}}.}

\item{input.data.tables}{List of array; each component is assumed to be a
Qtable and will be processed using.
\code{\link[flipTables]{AsTidyTabularData}}}

\item{input.data.raw}{List, containing variables or data.frames.}

\item{input.data.pasted}{List of length six; the first component of which is
assumed to be from a user-entered/pasted table; will be
processed by \code{\link{ParseUserEnteredTable}}.}

\item{input.data.other}{A PickAny Multi Q variable.}

\item{first.aggregate}{Logical; whether or not the input data
needs to be aggregated in this function. A single variable is tabulated,
2 variables are crosstabbed, and with 3 or more the mean is computed.
been provided, a contingency table is used to aggregate.}

\item{tidy}{Logical; whether or not the input data
needs to be aggregated in this function (e.g., if an x and y variable have
been provided, a contingency table is used to aggregate. This defaults
to \code{TRUE}. It aggressively seeks to turn the data into a named vector
or a matrix using \code{\link[flipTables]{TidyTabularData}}. This is not applied
when \code{data.input.tables} are provided, or when the chart type is
any of \code{"Scatter"}, \code{"Bean"}, \code{"Histogram"}, \code{"Density"},
\code{"Box"}, or \code{"Violin"}.}

\item{transpose}{Logical; should the resulting matrix (of created)  be
transposed?}

\item{missing}{Character; One of \code{"Error if missing data"},
\code{"Exclude cases with missing data"} (the default, which is
equivalent to 'complete.cases'), and \code{"Use partial data"},
which removes no data. Only used if either \code{input.data.raw} is provided,
or, a \code{weight} or \code{subset} are provided and are as long as the number of
observations in \code{data}.}

\item{row.names.to.remove}{Character vector or delimited string
of row labels specifying rows to remove from the returned table; default
is \code{c("NET", "SUM")}}

\item{column.names.to.remove}{Character vector or delimited string
of column labels specifying columns to remove from the returned table;
default is \code{c("NET", "SUM")}.}

\item{split}{Character delimiter to split \code{row.names.to.remove}
and \code{col.names.to.remove} on. Default is to split on either of
\code{","} or \code{";"}.  Assumed to be a regular expression; see
\code{\link{strsplit}}.}

\item{as.percentages}{Logical; If \code{TRUE}, aggregate values in the
output table are given as percentages summing to 100. If \code{FALSE},
column sums are given.}

\item{show.labels}{Logical; If \code{TRUE}, labels are used for
names in the data output if raw data is supplied.}
}
\value{
A list with components
\itemize{
\item \code{data} - If possible, a named vector or matrix, or if that is not
    posible or a data.frame is requested, a data.frame.
\item  \code{weights} - Numeric vector of user-supplied weights.
\item \code{y.title} - Character string to be used for the y-axis title; will
only be a non-empty string if some aggregation has been performed on
\code{data}
\item  \code{scatter.variable.indices} A named vector indicating which columns in
\code{data} should be plotted in a scatterplot as \code{x}, \code{y}, \code{sizes},
and \code{colors}. Is \code{NULL} if \code{chart.type} does not contain \code{"Scatter"}
or  \code{"Bubble"}. \code{NA} is used when the data does not exist.
}
}
\description{
Prepares input data for charting.
}
\details{
It is assumed that only one of \code{pasted},
    \code{input.data.table}, \code{input.data.tables}, \code{input.data.other},
    \code{input.data.raw} is non-NULL.  They are checked for nullity in
    that order.
}
\seealso{
\code{\link[flipTables]{AsTidyTabularData}},
    \code{\link[flipData]{TidyRawData}},
    \code{\link[flipTransformations]{ParseUserEnteredTable}}
}
