% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\docType{data}
\name{synonyms}
\alias{synonyms}
\title{Make sure that the first synonym is the briefiest, and from there they are ordered from
most more verbose to any roots.
These are swapped out in order. E.g., after the first line, all "colours" will be changed to 'col'.}
\format{An object of class \code{list} of length 8.}
\usage{
synonyms
}
\description{
Make sure that the first synonym is the briefiest, and from there they are ordered from
most more verbose to any roots.
These are swapped out in order. E.g., after the first line, all "colours" will be changed to 'col'.
}
\keyword{datasets}
